moves snake helpers to models_library
