+++ services/web/server/tests/unit/isolated/test_utils.py
@@ -14,7 +14,6 @@ from simcore_service_webserver.utils import (
     DATETIME_FORMAT,
     compute_sha1_on_small_dataset,
     now_str,
-    snake_to_camel,
     to_datetime,
 )
 from yarl import URL
@@ -34,21 +33,6 @@ def test_time_utils():
     now_time = datetime.utcnow()
     snapshot = now_time.strftime(DATETIME_FORMAT)
     assert now_time == datetime.strptime(snapshot, DATETIME_FORMAT)
-
-
-@pytest.mark.parametrize(
-    "subject,expected",
-    [
-        ("snAke_Fun", "snakeFun"),
-        ("", ""),
-        # since it assumes snake, notice how these cases get flatten
-        ("camelAlready", "camelalready"),
-        ("AlmostCamel", "almostcamel"),
-        ("_S", "S"),
-    ],
-)
-def test_snake_to_camel(subject, expected):
-    assert snake_to_camel(subject) == expected
 
 
 def test_yarl_url_compose_changed_with_latest_release():
